refactor: Pagination 컴포넌트 추가 및 ProblemTable에서 관련 코드 정리
