+++ src/entities/problems/ui/table/ProblemTable.tsx
@@ -1,16 +1,14 @@
 'use client';
-import { useState, useEffect, useMemo } from 'react';
+import { useState, useEffect } from 'react';
 import { useRouter } from 'next/navigation';
 import { ProblemsContent } from '@/entities/problems/model/types';
 import { useMyDailySolved } from '@/entities/mypage/model/query';
-import { CheckCircle, ChevronLeft, ChevronRight } from 'lucide-react';
-import { Button } from '@/components/ui/button';
+import { CheckCircle } from 'lucide-react';
 import Cookies from 'js-cookie';
 import { LevelUtil } from '@/shared/util/levelUtil';
 import TableHead from '@/entities/problems/ui/table/TableHead';
 import TableSkeleton from './TableSkeleton';
-
-const PAGE_LIMIT = 15;
+import Pagination from '@/widgets/pagination';
 
 export default function ProblemTable({
   data,
@@ -26,7 +24,6 @@ export default function ProblemTable({
   totalPages: number;
 }) {
   const router = useRouter();
-  const [pageGroupStart, setPageGroupStart] = useState<number>(0); // 0-based group start
   const [myProblemsList, setMyProblemsList] = useState<number[]>([]);
   const { data: myProblems } = useMyDailySolved();
   const token = Cookies.get('accessToken');
@@ -41,45 +38,6 @@ export default function ProblemTable({
     );
     setMyProblemsList(solved); // 하나의 배열에 담김
   }, [myProblems, token]);
-
-  useEffect(() => {
-    console.log(myProblemsList);
-  }, [myProblemsList]);
-
-  // currentPage가 page group 범위를 벗어나면 group start 재조정
-  useEffect(() => {
-    if (currentPage < pageGroupStart || currentPage >= pageGroupStart + PAGE_LIMIT) {
-      setPageGroupStart(Math.floor(currentPage / PAGE_LIMIT) * PAGE_LIMIT);
-    }
-  }, [currentPage, pageGroupStart]);
-
-  // totalPages는 "페이지 수" (count)라고 가정 -> lastIndex = totalPages - 1
-  const lastIndex = Math.max(totalPages - 1, 0);
-
-  const pageNumbers = useMemo(() => {
-    if (totalPages <= 0) return [];
-    const end = Math.min(pageGroupStart + PAGE_LIMIT - 1, lastIndex);
-    return Array.from({ length: end - pageGroupStart + 1 }, (_, i) => pageGroupStart + i);
-  }, [pageGroupStart, totalPages, lastIndex]);
-
-  // 오른쪽 화살표: 한 그룹 앞으로 (혹은 마지막 인덱스)
-  const handleNextPage = () => {
-    if (pageGroupStart + PAGE_LIMIT <= lastIndex) {
-      setCurrentPage(pageGroupStart + PAGE_LIMIT); // 다음 그룹 시작점
-    } else {
-      setCurrentPage(lastIndex); // 마지막 그룹에서 넘어가면 마지막 페이지로
-    }
-  };
-
-  // 왼쪽 화살표: 한 그룹 뒤로 (혹은 0)
-  const handlePrevPage = () => {
-    const prevGroupEnd = pageGroupStart - 1;
-    if (prevGroupEnd >= 0) {
-      setCurrentPage(prevGroupEnd); // 이전 그룹의 마지막 페이지
-    } else {
-      setCurrentPage(0); // 이미 첫 그룹이면 첫 페이지
-    }
-  };
 
   return (
     <div className="flex flex-col">
@@ -103,26 +61,24 @@ export default function ProblemTable({
                 return (
                   <tr
                     key={item.id}
-                    className={`border-b border-gray-800/50 hover:bg-white/[0.08] transition-colors duration-200 cursor-pointer ${
-                      isSolved ? 'bg-[#214d35]/10 border-l-4 border-l-[#00d084]' : ''
+                    className={`border-b border-gray-800/50 hover:bg-white/8 transition-colors duration-200 cursor-pointer ${
+                      isSolved ? 'bg-primary/10 border-l-4 border-l-secondary' : ''
                     }`}
                     onClick={() => router.push(`/problems/${item.id}`)}
                   >
                     <td className="px-6 py-4 text-sm text-center text-gray-300">{item.id}</td>
                     <td className="px-6 py-4">
                       <div className="flex flex-row gap-4 justify-center items-center">
-                        {isSolved && (
-                          <CheckCircle className="w-4 h-4 text-[#00d084] flex-shrink-0" />
-                        )}
+                        {isSolved && <CheckCircle className="w-4 h-4 text-secondary shrink-0" />}
                         <div>
-                          <div className="text-center text-sm font-medium text-white hover:text-[#00d084] transition-colors">
+                          <div className="text-center text-sm font-medium text-white hover:text-secondary transition-colors">
                             {item.title}
                           </div>
                           <div className="text-center text-xs text-gray-400 mt-1">{categories}</div>
                         </div>
                       </div>
                     </td>
-                    <td className="px-6 py-4 text-center text-sm text-[#00d084] font-medium">
+                    <td className="px-6 py-4 text-center text-sm text-secondary font-medium">
                       {item.score}
                     </td>
                     <td className={`px-6 py-4 text-center text-sm font-medium`}>
@@ -153,36 +109,12 @@ export default function ProblemTable({
         </table>
       </div>
 
-      {/* 페이지네이션 */}
-      <div className="flex justify-center items-center gap-2 mt-4 text-gray-400">
-        <Button
-          onClick={handlePrevPage}
-          className="p-1 bg-[#214d35] hover:bg-[#276e48] disabled:opacity-50 disabled:cursor-not-allowed rounded-[10px] w-10 h-10"
-        >
-          <ChevronLeft className="w-4 h-4" />
-        </Button>
-        {pageNumbers.map((num) => (
-          <Button
-            key={num}
-            onClick={() => {
-              if (!isLoading) setCurrentPage(num);
-            }}
-            className={`w-10 h-10 rounded-md flex items-center justify-center text-sm transition ${
-              currentPage === num
-                ? 'bg-[#214d35] hover:bg-[#276e48] text-white border-[#214d35]'
-                : 'bg-gray-800 border-gray-700 text-white hover:bg-gray-700'
-            }`}
-          >
-            {num + 1}
-          </Button>
-        ))}
-        <Button
-          onClick={handleNextPage}
-          className="p-1 bg-[#214d35] hover:bg-[#276e48] disabled:opacity-50 disabled:cursor-not-allowed rounded-[10px] w-10 h-10"
-        >
-          <ChevronRight className="w-4 h-4" />
-        </Button>
-      </div>
+      <Pagination
+        currentPage={currentPage}
+        setCurrentPage={setCurrentPage}
+        totalPages={totalPages}
+        isLoading={isLoading}
+      />
     </div>
   );
 }
